requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
twilio==8.13.0
openai==1.3.7
python-multipart==0.0.6
jinja2==3.1.2